requirements.txt
aiogram>=3.15,<4
openai>=1.60,<2
psycopg[binary]>=3.2,<4
pyyaml>=6.0
python-dotenv>=1.0

